Data_project_2/requirements.txt
geopandas==0.14.2
folium==0.15.1
simplekml==1.3.6
shapely==2.0.2
streamlit_folium==0.18.0
faker==21.0.0
argparse==1.4.0
google-cloud-pubsub==2.18.1
google-api-core==1.34.0
proto-plus==1.23.0
apache-beam==2.51.0
google-apitools==0.5.32
google-cloud-bigquery==3.13.0
pandas==1.5.3
protobuf==3.20.3
